requirements.txt
aiogram==3.13.1
aiohttp
python-dotenv